+++ example/lib/main.dart
@@ -2,11 +2,17 @@ import 'dart:convert';
 
 import 'package:example/flavors/main_dev.dart';
 import 'package:example/network/network_manager.dart';
+import 'package:example/res/res.dart';
+import 'package:example/res/strings/main/bg_strings.dart';
+import 'package:example/res/strings/main/en_strings.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:g_base_package/base/app_exception.dart';
+import 'package:g_base_package/base/ui/logs_screen.dart';
 import 'package:g_base_package/base/utils/dialogs.dart';
 import 'package:g_base_package/base/utils/logger.dart';
 import 'package:g_base_package/base/utils/system.dart';
+import 'package:g_base_package/base/utils/utils.dart';
 import 'package:g_base_package/base/utils/versions.dart';
 import 'package:g_base_package/base/utils/files.dart';
 import 'package:image_picker/image_picker.dart';
@@ -16,6 +22,8 @@ import 'model/session.dart';
 
 void main() {
   DevConfig();
+  Localization.init(null, [EnUSStrings(), BgStrings()], EnUSStrings(),
+      globalLocales: [EnUSGlobalStrings(), BgGlobalStrings()]);
   runApp(MyApp());
 }
 
@@ -38,7 +46,9 @@ class MyApp extends StatelessWidget {
         // is not restarted.
         primarySwatch: Colors.blue,
       ),
-      home: MyHomePage(title: 'Flutter Demo Home Page'),
+      home: MyHomePage(
+        title: Txt.get(StrKey.mainTitle),
+      ),
     );
   }
 }
@@ -139,6 +149,9 @@ class _MyHomePageState extends State<MyHomePage> {
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
             Text(
+              Txt.get(StrKey.appName),
+            ),
+            Text(
               'You have pushed the button this many times:',
             ),
             Text(
@@ -203,12 +216,7 @@ class _MyHomePageState extends State<MyHomePage> {
                 );
               },
             ),
-            FlatButton(
-              child: Text("Update"),
-              onPressed: () {
-                NetworkManager(token).updateWorkspace();
-              },
-            ),
+
             FlatButton(
               child: Text("Select and Upload"),
               onPressed: () {
@@ -242,12 +250,16 @@ class _MyHomePageState extends State<MyHomePage> {
       });
 
       if (copiedFile != null && await copiedFile.exists()) {
-        NetworkManager(token).uploadImage(
+        NetworkManager(token).uploadImageFuture(
           copiedFile.path,
-          companyId,
-          "9e625fb6-e81a-414d-bcfe-95c9d8d80001",
-          (_) {},
-        );
+          "b55306bc-20d0-4ee6-adb1-d3307c308502",
+          "2ac7d50d-da40-41a8-b84d-3a87c0fb9e4a",
+          (json) {
+            Log.d(json, tag);
+          },
+        ).catchError((e) {
+          Log.error("uploadImageFuture", error: e);
+        });
       }
     }
   }